refactor: Renamed components not to use component word
