+++ src/components/navigation/Navbar.module.scss
@@ -1,0 +1,5 @@
+.header {
+  width: 100%;
+  position: sticky;
+  top: 0px;
+}
