--- src/components/navigation/NavbarComponent.module.scss
@@ -1,5 +1,0 @@
-.header {
-  width: 100%;
-  position: sticky;
-  top: 0px;
-}
