+++ src/components/profile/Profile.module.scss
@@ -1,0 +1,4 @@
+.profileContainer {
+  margin: 10px;
+  background-color: white;
+}
