--- src/components/profile/ProfileComponent.module.scss
@@ -1,4 +1,0 @@
-.profileContainer {
-  margin: 10px;
-  background-color: white;
-}
